refactor(#8): 디렉토리 이름 변경
Avatar -> ProfileImage
